+++ src/main/java/com/yanny/ytech/compatibility/JeiCompatibility.java
@@ -29,6 +29,7 @@ public class JeiCompatibility implements IModPlugin {
         registration.addRecipeCategories(new BlockHitRecipeCategory(registration.getJeiHelpers().getGuiHelper()));
         registration.addRecipeCategories(new AlloyingRecipeCategory(registration.getJeiHelpers().getGuiHelper()));
         registration.addRecipeCategories(new HammeringRecipeCategory(registration.getJeiHelpers().getGuiHelper()));
+        registration.addRecipeCategories(new PotteryRecipeCategory(registration.getJeiHelpers().getGuiHelper()));
     }
 
     @Override
@@ -43,6 +44,7 @@ public class JeiCompatibility implements IModPlugin {
             registration.addRecipes(BlockHitRecipeCategory.RECIPE_TYPE, BlockHitRecipeCategory.getRecipes(level.getRecipeManager()));
             registration.addRecipes(AlloyingRecipeCategory.RECIPE_TYPE, AlloyingRecipeCategory.getRecipes(level.getRecipeManager()));
             registration.addRecipes(HammeringRecipeCategory.RECIPE_TYPE, HammeringRecipeCategory.getRecipes(level.getRecipeManager()));
+            registration.addRecipes(PotteryRecipeCategory.RECIPE_TYPE, PotteryRecipeCategory.getRecipes(level.getRecipeManager()));
 
             registration.addItemStackInfo(YTechItems.GRASS_FIBERS.get().getDefaultInstance(), Component.translatable("text.ytech.info.grass_fibers"));
         } else {
@@ -59,6 +61,7 @@ public class JeiCompatibility implements IModPlugin {
         BlockHitRecipeCategory.registerCatalyst(registration);
         AlloyingRecipeCategory.registerCatalyst(registration);
         HammeringRecipeCategory.registerCatalyst(registration);
+        PotteryRecipeCategory.registerCatalyst(registration);
     }
 
     @NotNull