+++ src/main/java/com/yanny/ytech/registration/YTechBlocks.java
@@ -27,6 +27,7 @@ public class YTechBlocks {
     public static final RegistryObject<Block> FIRE_PIT = BLOCKS.register("fire_pit", FirePitBlock::new);
     public static final RegistryObject<Block> GRASS_BED = BLOCKS.register("grass_bed", GrassBedBlock::new);
     public static final RegistryObject<Block> MILLSTONE = BLOCKS.register("millstone", MillstoneBlock::new);
+    public static final RegistryObject<Block> POTTERS_WHEEL = BLOCKS.register("potters_wheel", PottersWheelBlock::new);
     public static final RegistryObject<Block> PRIMITIVE_ALLOY_SMELTER = BLOCKS.register("primitive_alloy_smelter", PrimitiveAlloySmelterBlock::new);
     public static final RegistryObject<Block> PRIMITIVE_SMELTER = BLOCKS.register("primitive_smelter", PrimitiveSmelterBlock::new);
     public static final RegistryObject<Block> REINFORCED_BRICKS = registerBlock("reinforced_bricks", Blocks.BRICKS);