+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # YTech Changelog
 
+## []
+
+- Added Potter's Wheel
+- Changed Unfired Clay Bucket recipe
+
 ## [0.3.3]
 
 - Fixed broken recipes for Thatch and Stone Hammer and changed recipe for grass bed